feat: add unified interface for different modes of convolution
and removed mutex from struct arg
